+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/table/DefaultFilterFieldFactory.java
@@ -19,8 +19,6 @@
 package is.codion.swing.common.ui.component.table;
 
 import is.codion.common.model.condition.ConditionModel;
-import is.codion.common.value.Value;
-import is.codion.common.value.ValueSet;
 import is.codion.swing.common.ui.component.builder.ComponentBuilder;
 import is.codion.swing.common.ui.component.table.FilterColumnConditionPanel.FieldFactory;
 
@@ -51,43 +49,41 @@ final class DefaultFilterFieldFactory<C> implements FieldFactory<C> {
 	}
 
 	@Override
-	public JComponent createEqualField(ConditionModel<?> condition, C identifier) {
+	public <T> JComponent createEqualField(ConditionModel<T> condition, C identifier) {
 		return createField(condition)
-						.link((Value<Object>) condition.operands().equal())
+						.link(condition.operands().equal())
 						.build();
 	}
 
 	@Override
-	public Optional<JComponent> createUpperBoundField(ConditionModel<?> condition, C identifier) {
+	public <T> Optional<JComponent> createUpperBoundField(ConditionModel<T> condition, C identifier) {
 		if (condition.valueClass().equals(Boolean.class)) {
 			return Optional.empty();//no upper bound field required for boolean values
 		}
 
 		return Optional.of(createField(condition)
-						.link((Value<Object>) condition.operands().upperBound())
+						.link(condition.operands().upperBound())
 						.build());
 	}
 
 	@Override
-	public Optional<JComponent> createLowerBoundField(ConditionModel<?> condition, C identifier) {
+	public <T> Optional<JComponent> createLowerBoundField(ConditionModel<T> condition, C identifier) {
 		if (condition.valueClass().equals(Boolean.class)) {
 			return Optional.empty();//no lower bound field required for boolean values
 		}
 
 		return Optional.of(createField(condition)
-						.link((Value<Object>) condition.operands().lowerBound())
+						.link(condition.operands().lowerBound())
 						.build());
 	}
 
 	@Override
-	public JComponent createInField(ConditionModel<?> condition, C identifier) {
-		return listBox(createField(condition).buildValue(),
-						(ValueSet<Object>) condition.operands().in())
-						.build();
+	public <T> JComponent createInField(ConditionModel<T> condition, C identifier) {
+		return listBox(createField(condition).buildValue(), condition.operands().in()).build();
 	}
 
-	private static <T> ComponentBuilder<T, ? extends JComponent, ?> createField(ConditionModel<?> conditionModel) {
-		Class<?> columnClass = conditionModel.valueClass();
+	private static <T> ComponentBuilder<T, ? extends JComponent, ?> createField(ConditionModel<T> conditionModel) {
+		Class<T> columnClass = conditionModel.valueClass();
 		if (columnClass.equals(Boolean.class)) {
 			return (ComponentBuilder<T, ? extends JComponent, ?>) checkBox()
 							.nullable(true)