+++ changelog.md
@@ -30,6 +30,7 @@ Codion Change Log
 - FilterTableCellRenderer.Builder.condition() renamed filter().
 - TableConditionPanel.conditionModel() renamed conditions().
 - FilterTable.Builder.resizeRowToFitEditor() added along with RESIZE_ROW_TO_FIT_EDITOR configuration value.
+- FilterColumnConditionPanel.FieldFactory, type parameter added to create methods.
 ### is.codion.framework.model
 - DefaultEntityConditionModel, no longer wraps a single condition in a condition combination.
 - EntityConditionModel.AdditionalCondition added, related refactoring.